+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* An `IsPremultipliedAlpha` property to the `HeifImage` and `HeifImageHandle` classes.
+  * This property is supported on LibHeif 1.12 and later.
+
 ## v2.1.0 - 2021-05-19
 
 ### Deprecated